@@ -1,6 +1,7 @@
 [build-system]
 requires = [
-    "setuptools",
+    "setuptools >= 77",
+    "setuptools_scm >= 8",
     "wheel",
 ]
 build-backend = "setuptools.build_meta"
@@ -19,7 +20,7 @@ authors = [
     {name = "Michele Ceriotti"}
 ]
 readme = "README.rst"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = {text = "BSD-3-Clause"}
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -31,20 +32,20 @@ classifiers = [
     "Operating System :: MacOS :: MacOS X",
     "Operating System :: Microsoft :: Windows",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering",
 ]
 dependencies = [
-    "scikit-learn>=1.1.0",
+    "scikit-learn >= 1.7, < 1.8",
+    "scipy >= 1.15",  # explicit to adhere to scikit-learn dependencies
 ]
 dynamic = ["version"]
 
 [project.optional-dependencies]
 examples = [
-    "ase",
     "matplotlib",
     "pandas",
     "tqdm",
@@ -57,11 +58,17 @@ repository = "https://github.com/scikit-learn-contrib/scikit-matter"
 issues = "https://github.com/scikit-learn-contrib/scikit-matterissues"
 changelog = "http://scikit-matter.readthedocs.io/en/latest/changelog.html"
 
+[tool.check-manifest]
+ignore = ["src/skmatter/_version.py"]
+
 [tool.setuptools.packages.find]
 where = ["src"]
 
 [tool.setuptools.dynamic]
 version = {attr = "skmatter.__version__"}
+
+[tool.setuptools_scm]
+version_file = "src/skmatter/_version.py"
 
 [tool.coverage.run]
 branch = true
@@ -76,13 +83,39 @@ include = [
 output = 'tests/coverage.xml'
 
 [tool.pytest.ini_options]
-testpaths = "tests"
+testpaths = ["tests"]
+addopts = [
+    "--cov",
+    "--cov-append",
+    "--cov-report=",
+    "--import-mode=append",
+]
 
-[tool.isort]
-skip = "__init__.py"
-profile = "black"
-line_length = 88
-indent = 4
-include_trailing_comma = true
-lines_after_imports = 2
-known_first_party = "skmatter"
+[tool.ruff]
+exclude = ["docs/src/examples/", "src/torchpme/_version.py"]
+lint.ignore = [
+    "F401",
+    "E203",
+    "D100",
+    "D101",
+    "D102",
+    "D205",
+    "D400",
+    "D401",
+]
+line-length = 88
+lint.select = [
+    "D",
+    "E",
+    "F",
+    "W",
+]
+
+[tool.ruff.lint.pydocstyle]
+convention = "numpy"
+
+[tool.ruff.lint.per-file-ignores]
+"examples/**" = [
+    "D205",
+    "D400",
+]
