@@ -1,5 +1,6 @@
 import unittest
 
+import numpy as np
 from sklearn.datasets import load_diabetes as get_dataset
 from sklearn.utils.validation import NotFittedError
 
@@ -13,7 +14,7 @@ class TestFPS(unittest.TestCase):
 
     def test_restart(self):
         """
-        This test checks that the model can be restarted with a new number of
+        Check that the model can be restarted with a new number of
         features and `warm_start`
         """
         selector = FPS(n_to_select=1, initialize=self.idx[0])
@@ -25,11 +26,9 @@ class TestFPS(unittest.TestCase):
             self.assertEqual(selector.selected_idx_[i - 1], self.idx[i - 1])
 
     def test_initialize(self):
+        """Check that the model can be initialized in all applicable manners and throws
+        an error otherwise.
         """
-        This test checks that the model can be initialized in all applicable manners
-        and throws an error otherwise
-        """
-
         for initialize in [self.idx[0], "random"]:
             with self.subTest(initialize=initialize):
                 selector = FPS(n_to_select=1, initialize=initialize)
@@ -42,15 +41,38 @@ class TestFPS(unittest.TestCase):
             for i in range(4):
                 self.assertEqual(selector.selected_idx_[i], self.idx[i])
 
+        initialize = np.array(self.idx[:4])
+        with self.subTest(initialize=initialize):
+            selector = FPS(n_to_select=len(self.idx) - 1, initialize=initialize)
+            selector.fit(self.X)
+            for i in range(4):
+                self.assertEqual(selector.selected_idx_[i], self.idx[i])
+
+        initialize = np.array([1, 5, 3, 0.25])
+        with self.subTest(initialize=initialize):
+            with self.assertRaises(ValueError) as cm:
+                selector = FPS(n_to_select=len(self.idx) - 1, initialize=initialize)
+                selector.fit(self.X)
+            self.assertEqual(
+                str(cm.exception), "Invalid value of the initialize parameter"
+            )
+
+        initialize = np.array([[1, 5, 3], [2, 4, 6]])
+        with self.subTest(initialize=initialize):
+            with self.assertRaises(ValueError) as cm:
+                selector = FPS(n_to_select=len(self.idx) - 1, initialize=initialize)
+                selector.fit(self.X)
+            self.assertEqual(
+                str(cm.exception), "Invalid value of the initialize parameter"
+            )
+
         with self.assertRaises(ValueError) as cm:
             selector = FPS(n_to_select=1, initialize="bad")
             selector.fit(self.X)
         self.assertEqual(str(cm.exception), "Invalid value of the initialize parameter")
 
     def test_get_distances(self):
-        """
-        This test checks that the hausdorff distances are returnable after fitting
-        """
+        """Check that the hausdorff distances are returnable after fitting."""
         selector = FPS(n_to_select=7)
         selector.fit(self.X)
         d = selector.get_select_distance()
@@ -61,6 +83,22 @@ class TestFPS(unittest.TestCase):
         with self.assertRaises(NotFittedError):
             selector = FPS(n_to_select=7)
             _ = selector.get_select_distance()
+
+    def test_unique_selected_idx_zero_score(self):
+        """
+        Tests that the selected idxs are unique, which may not be the
+        case when the score is numerically zero
+        """
+        np.random.seed(0)
+        n_samples = 10
+        n_features = 15
+        X = np.random.rand(n_samples, n_features)
+        X[:, 1] = X[:, 0]
+        X[:, 2] = X[:, 0]
+        selector_problem = FPS(n_to_select=len(X.T)).fit(X)
+        assert len(selector_problem.selected_idx_) == len(
+            set(selector_problem.selected_idx_)
+        )
 
 
 if __name__ == "__main__":
