@@ -11,6 +11,8 @@ from skmatter.metrics import (
     global_reconstruction_error,
     local_prediction_rigidity,
     local_reconstruction_error,
+    pairwise_mahalanobis_distances,
+    periodic_pairwise_euclidean_distances,
     pointwise_local_reconstruction_error,
 )
 
@@ -47,8 +49,7 @@ class PredictionRigidityTests(unittest.TestCase):
         )
         self.assertTrue(
             rank_diff == 0,
-            f"LPR Covariance matrix rank is not full, with a difference of:"
-            f"{rank_diff}",
+            f"LPR Covariance matrix rank is not full, with a difference of:{rank_diff}",
         )
 
     def test_componentwise_prediction_rigidity(self):
@@ -211,6 +212,65 @@ class ReconstructionMeasuresTests(unittest.TestCase):
             len(lfre_val) == test_size,
             f"size of pointwise LFRE  {len(lfre_val)} differs from expected test set "
             f"size {test_size}",
+        )
+
+
+class DistanceTests(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.X = np.array([[1, 2], [3, 4], [5, 6]])
+        cls.Y = np.array([[7, 8], [9, 10]])
+        cls.covs = np.array([[[1, 0.5], [0.5, 1]], [[1, 0.0], [0.0, 1]]])
+        cls.cell = [5, 7]
+        cls.distances = np.array(
+            [
+                [8.48528137, 11.3137085],
+                [5.65685425, 8.48528137],
+                [2.82842712, 5.65685425],
+            ]
+        )
+        cls.periodic_distances = np.array(
+            [
+                [1.41421356, 2.23606798],
+                [3.16227766, 1.41421356],
+                [2.82842712, 3.16227766],
+            ]
+        )
+        cls.mahalanobis_distances = np.array(
+            [
+                [
+                    [10.39230485, 13.85640646],
+                    [6.92820323, 10.39230485],
+                    [3.46410162, 6.92820323],
+                ],
+                cls.distances,
+            ]
+        )
+
+    def test_euclidean_distance(self):
+        distances = periodic_pairwise_euclidean_distances(self.X, self.Y)
+        self.assertTrue(
+            np.allclose(distances, self.distances),
+            f"Calculated distance does not match expected value"
+            f"Calculated: {distances} Expected: {self.distances}",
+        )
+
+    def test_periodic_euclidean_distance(self):
+        distances = periodic_pairwise_euclidean_distances(
+            self.X, self.Y, cell_length=self.cell
+        )
+        self.assertTrue(
+            np.allclose(distances, self.periodic_distances),
+            f"Calculated distance does not match expected value"
+            f"Calculated: {distances} Expected: {self.periodic_distances}",
+        )
+
+    def test_mahalanobis_distance(self):
+        distances = pairwise_mahalanobis_distances(self.X, self.Y, self.covs)
+        self.assertTrue(
+            np.allclose(distances, self.mahalanobis_distances),
+            f"Calculated distance does not match expected value"
+            f"Calculated: {distances} Expected: {self.mahalanobis_distances}",
         )
 
 
